+++ src/de/deserializer.rs
@@ -37,7 +37,7 @@
 use core::convert::TryInto;
 use serde::de::{DeserializeSeed, IntoDeserializer, Visitor};
 
-use crate::varint::read_mqtt_u32_varint;
+use crate::{trace, varint::read_mqtt_u32_varint};
 
 /// Errors returned while decoding MQTT packets.
 #[derive(Debug, Copy, Clone, PartialEq)]
@@ -64,7 +64,7 @@ impl serde::ser::StdError for Error {}
 
 impl serde::de::Error for Error {
     fn custom<T: core::fmt::Display>(_msg: T) -> Self {
-        crate::trace!("Deserialization error");
+        trace!("Deserialization error");
         Error::Custom
     }
 }
@@ -103,12 +103,12 @@ impl<'a> MqttDeserializer<'a> {
     }
 
     /// Override the next binary field length.
-    pub(crate) fn set_next_bytes_len(&mut self, len: usize) {
+    fn set_next_bytes_len(&mut self, len: usize) {
         self.next_bytes_len.replace(len);
     }
 
     /// Attempt to take N bytes from the buffer.
-    pub(crate) fn try_take_n(&mut self, n: usize) -> Result<&'a [u8], Error> {
+    fn try_take_n(&mut self, n: usize) -> Result<&'a [u8], Error> {
         if self.len() < n {
             return Err(Error::InsufficientData);
         }
@@ -119,7 +119,7 @@ impl<'a> MqttDeserializer<'a> {
     }
 
     /// Pop a single byte from the data buffer.
-    pub(crate) fn pop(&mut self) -> Result<u8, Error> {
+    fn pop(&mut self) -> Result<u8, Error> {
         if self.len() == 0 {
             return Err(Error::InsufficientData);
         }
@@ -130,17 +130,17 @@ impl<'a> MqttDeserializer<'a> {
     }
 
     /// Read a 16-bit integer from the data buffer.
-    pub(crate) fn read_u16(&mut self) -> Result<u16, Error> {
+    fn read_u16(&mut self) -> Result<u16, Error> {
         Ok(u16::from_be_bytes([self.pop()?, self.pop()?]))
     }
 
     /// Read the number of remaining bytes in the data buffer.
-    pub(crate) fn len(&self) -> usize {
+    fn len(&self) -> usize {
         self.buf.len() - self.index
     }
 
     /// Read a variable-length integer from the data buffer.
-    pub(crate) fn read_varint(&mut self) -> Result<u32, Error> {
+    fn read_varint(&mut self) -> Result<u32, Error> {
         read_mqtt_u32_varint(|| self.pop(), || Error::BadVarint)
     }
 
@@ -440,7 +440,7 @@ impl<'de> serde::de::EnumAccess<'de> for &'_ mut MqttDeserializer<'de> {
 
     fn variant_seed<V: DeserializeSeed<'de>>(self, seed: V) -> Result<(V::Value, Self), Error> {
         let varint = self.read_varint()?;
-        crate::trace!("Read Varint: {=u32:#X}", varint);
+        trace!("Read Varint: {=u32:#X}", varint);
         let v = DeserializeSeed::deserialize(seed, varint.into_deserializer())?;
         Ok((v, self))
     }