+++ src/de/mod.rs
@@ -1,5 +1,6 @@
-pub(crate) mod deserializer;
+mod deserializer;
 mod packet_reader;
-pub(crate) mod received_packet;
-pub(crate) use deserializer::Error;
+mod received_packet;
+pub(crate) use deserializer::{Error, MqttDeserializer};
 pub(crate) use packet_reader::PacketReader;
+pub(crate) use received_packet::ReceivedPacket;