+++ src/packets.rs
@@ -14,7 +14,6 @@ pub(crate) struct Connect<'a> {
     pub(crate) keepalive: u16,
     pub(crate) properties: Properties<'a>,
     pub(crate) client_id: Utf8String<'a>,
-    #[allow(dead_code)]
     pub(crate) auth: Option<Auth<'a>>,
     pub(crate) will: Option<Will<'a>>,
     pub(crate) clean_start: bool,
@@ -151,10 +150,6 @@ impl serde::Serialize for TopicFilters<'_> {
 #[derive(Debug, Serialize)]
 pub(crate) struct PingReq;
 
-#[allow(dead_code)]
-#[derive(Debug, Deserialize)]
-pub(crate) struct PingResp;
-
 #[derive(Debug, Deserialize, Serialize)]
 pub(crate) struct PubAck<'a> {
     pub(crate) packet_id: u16,
@@ -162,22 +157,20 @@ pub(crate) struct PubAck<'a> {
     pub(crate) reason: Reason<'a>,
 }
 
-#[allow(dead_code)]
 #[derive(Debug, Deserialize)]
 pub(crate) struct SubAck<'a> {
     pub(crate) packet_id: u16,
-    #[serde(borrow)]
-    pub(crate) properties: Properties<'a>,
+    #[serde(borrow, rename = "properties")]
+    pub(crate) _properties: Properties<'a>,
     #[serde(skip)]
     pub(crate) codes: &'a [u8],
 }
 
-#[allow(dead_code)]
 #[derive(Debug, Deserialize)]
 pub(crate) struct UnsubAck<'a> {
     pub(crate) packet_id: u16,
-    #[serde(borrow)]
-    pub(crate) properties: Properties<'a>,
+    #[serde(borrow, rename = "properties")]
+    pub(crate) _properties: Properties<'a>,
     #[serde(skip)]
     pub(crate) codes: &'a [u8],
 }
@@ -289,7 +282,7 @@ impl From<ReasonCode> for Reason<'_> {
         Self {
             reason: Some(ReasonData {
                 code,
-                properties: None,
+                _properties: None,
             }),
         }
     }
@@ -297,14 +290,11 @@ impl From<ReasonCode> for Reason<'_> {
 
 impl<'a> Reason<'a> {
     #[cfg_attr(not(feature = "fuzzing"), allow(dead_code))]
-    pub(crate) fn with_properties(
-        code: ReasonCode,
-        properties: &'a [crate::properties::Property<'a>],
-    ) -> Self {
+    pub(crate) fn with_properties(code: ReasonCode, properties: &'a [Property<'a>]) -> Self {
         Self {
             reason: Some(ReasonData {
                 code,
-                properties: Some(Properties::from_slice(properties)),
+                _properties: Some(Properties::from_slice(properties)),
             }),
         }
     }
@@ -320,17 +310,20 @@ impl<'a> Reason<'a> {
 #[derive(Debug, Deserialize, Serialize)]
 struct ReasonData<'a> {
     pub code: ReasonCode,
-    #[serde(borrow)]
-    #[allow(dead_code)]
-    pub properties: Option<Properties<'a>>,
+    #[serde(borrow, rename = "properties")]
+    pub _properties: Option<Properties<'a>>,
 }
 
 #[cfg(test)]
 mod tests {
-    use super::Disconnect;
+    use super::{
+        Connect, Disconnect, PingReq, PubAck, PubComp, PubRel, PublishHeader, Reason, ReasonData,
+        Subscribe, Unsubscribe,
+    };
     use crate::properties::{Properties, Property};
     use crate::reason_codes::ReasonCode;
     use crate::ser::MqttSerializer;
+    use crate::will::Will;
     use crate::{QoS, Retain};
     use crate::{types::TopicFilter, wire::Utf8String};
 
@@ -344,7 +337,7 @@ mod tests {
             0xAB, 0xCD, // Payload
         ];
 
-        let header = crate::packets::PublishHeader {
+        let header = PublishHeader {
             qos: QoS::AtMostOnce,
             packet_id: None,
             dup: false,
@@ -371,7 +364,7 @@ mod tests {
             0xAB, 0xCD, // Payload
         ];
 
-        let header = crate::packets::PublishHeader {
+        let header = PublishHeader {
             qos: QoS::AtLeastOnce,
             topic: Utf8String("ABC"),
             packet_id: Some(0xBEEF),
@@ -398,7 +391,7 @@ mod tests {
             0xAB, 0xCD, // Payload
         ];
 
-        let header = crate::packets::PublishHeader {
+        let header = PublishHeader {
             qos: QoS::AtLeastOnce,
             topic: Utf8String("ABC"),
             packet_id: Some(0xBEEF),
@@ -425,7 +418,7 @@ mod tests {
             0x00, // Options byte = 0
         ];
 
-        let subscribe = crate::packets::Subscribe {
+        let subscribe = Subscribe {
             packet_id: 16,
             dup: false,
             properties: Properties::from_slice(&[]),
@@ -448,7 +441,7 @@ mod tests {
             0x00, 0x03, 0x41, 0x42, 0x43, // Topic: ABC
         ];
 
-        let unsubscribe = crate::packets::Unsubscribe {
+        let unsubscribe = Unsubscribe {
             packet_id: 16,
             dup: false,
             properties: Properties::from_slice(&[]),
@@ -472,7 +465,7 @@ mod tests {
             0xAB, 0xCD, // Payload
         ];
 
-        let header = crate::packets::PublishHeader {
+        let header = PublishHeader {
             qos: QoS::AtMostOnce,
             topic: Utf8String("ABC"),
             packet_id: None,
@@ -499,7 +492,7 @@ mod tests {
             0xAB, 0xCD, // Payload
         ];
 
-        let header = crate::packets::PublishHeader {
+        let header = PublishHeader {
             qos: QoS::AtMostOnce,
             topic: Utf8String("ABC"),
             packet_id: None,
@@ -528,7 +521,7 @@ mod tests {
         ];
 
         let mut buffer: [u8; 900] = [0; 900];
-        let connect: crate::packets::Connect<'_> = crate::packets::Connect {
+        let connect: Connect<'_> = Connect {
             client_id: Utf8String("ABC"),
             auth: None,
             will: None,
@@ -571,11 +564,11 @@ mod tests {
         ];
 
         let mut buffer: [u8; 900] = [0; 900];
-        let will = crate::will::Will::new("EFG", &[0xAB, 0xCD], &[])
+        let will = Will::new("EFG", &[0xAB, 0xCD], &[])
             .unwrap()
-            .qos(crate::QoS::AtMostOnce);
+            .qos(QoS::AtMostOnce);
 
-        let connect = crate::packets::Connect {
+        let connect = Connect {
             clean_start: true,
             keepalive: 10,
             properties: Properties::from_slice(&[]),
@@ -597,7 +590,7 @@ mod tests {
         ];
 
         let mut buffer: [u8; 1024] = [0; 1024];
-        let message = MqttSerializer::encode(&mut buffer, &crate::packets::PingReq {}).unwrap();
+        let message = MqttSerializer::encode(&mut buffer, &PingReq {}).unwrap();
         assert_eq!(message, good_ping_req);
     }
 
@@ -633,7 +626,7 @@ mod tests {
             0x10, // Response Code
         ];
 
-        let pubrel = crate::packets::PubRel {
+        let pubrel = PubRel {
             packet_id: 5,
             reason: ReasonCode::NoMatchingSubscribers.into(),
         };
@@ -655,12 +648,12 @@ mod tests {
             0x00, // Response Code
         ];
 
-        let pubrel = crate::packets::PubAck {
+        let pubrel = PubAck {
             packet_id: 0x15,
-            reason: crate::packets::Reason {
-                reason: Some(crate::packets::ReasonData {
+            reason: Reason {
+                reason: Some(ReasonData {
                     code: ReasonCode::Success,
-                    properties: None,
+                    _properties: None,
                 }),
             },
         };
@@ -682,12 +675,12 @@ mod tests {
             0x00, // Response Code
         ];
 
-        let pubrel = crate::packets::PubComp {
+        let pubrel = PubComp {
             packet_id: 0x15,
-            reason: crate::packets::Reason {
-                reason: Some(crate::packets::ReasonData {
+            reason: Reason {
+                reason: Some(ReasonData {
                     code: ReasonCode::Success,
-                    properties: None,
+                    _properties: None,
                 }),
             },
         };